chore: move generate username to user service
