+++ models/users.js
@@ -791,26 +791,6 @@ const getUsersByRole = async (role) => {
   }
 };
 
-const generateUniqueUsername = async (firstName, lastName) => {
-  try {
-    const snapshot = await userModel
-      .where("first_name", "==", firstName)
-      .where("last_name", "==", lastName)
-      .count()
-      .get();
-
-    const existingUserCount = snapshot.data().count || 0;
-
-    const suffix = existingUserCount + 1;
-    const finalUsername = formatUsername(firstName, lastName, suffix);
-
-    return finalUsername;
-  } catch (err) {
-    logger.error(`Error while generating unique username: ${err.message}`);
-    throw err;
-  }
-};
-
 /**
  * Updates given list of users in batch
  * @param usersData {Array} - Users list as an array.
@@ -1049,7 +1029,6 @@ module.exports = {
   removeGitHubToken,
   getUsersByRole,
   fetchUserByIds,
-  generateUniqueUsername,
   updateUsersInBatch,
   fetchUsersListForMultipleValues,
   fetchUserForKeyValue,
